Move get_block_template() support functions to a submodule
