+++ zebra-rpc/src/methods/get_block_template_rpcs.rs
@@ -1,6 +1,6 @@
 //! RPC methods related to mining only available with `getblocktemplate-rpcs` rust feature.
 
-use std::{iter, sync::Arc};
+use std::sync::Arc;
 
 use futures::{FutureExt, TryFutureExt};
 use jsonrpc_core::{self, BoxFuture, Error, ErrorCode, Result};
@@ -8,27 +8,19 @@ use jsonrpc_derive::rpc;
 use tower::{buffer::Buffer, Service, ServiceExt};
 
 use zebra_chain::{
-    amount::{self, Amount, NegativeOrZero, NonNegative},
     block::{
-        self,
-        merkle::{self, AuthDataRoot},
-        Block, ChainHistoryBlockTxAuthCommitmentHash, Height, MAX_BLOCK_BYTES, ZCASH_BLOCK_VERSION,
+        self, Block, ChainHistoryBlockTxAuthCommitmentHash, Height, MAX_BLOCK_BYTES,
+        ZCASH_BLOCK_VERSION,
     },
     chain_sync_status::ChainSyncStatus,
     chain_tip::ChainTip,
     parameters::Network,
     serialization::ZcashDeserializeInto,
-    transaction::{Transaction, UnminedTx, VerifiedUnminedTx},
+    transaction::Transaction,
     transparent,
 };
-
-use zebra_consensus::{
-    funding_stream_address, funding_stream_values, miner_subsidy, VerifyChainError,
-    MAX_BLOCK_SIGOPS,
-};
-
+use zebra_consensus::{VerifyChainError, MAX_BLOCK_SIGOPS};
 use zebra_node_services::mempool;
-
 use zebra_state::{ReadRequest, ReadResponse};
 
 use crate::methods::{
@@ -38,6 +30,10 @@ use crate::methods::{
             DEFAULT_SOLUTION_RATE_WINDOW_SIZE, GET_BLOCK_TEMPLATE_CAPABILITIES_FIELD,
             GET_BLOCK_TEMPLATE_MUTABLE_FIELD, GET_BLOCK_TEMPLATE_NONCE_RANGE_FIELD,
             MAX_ESTIMATED_DISTANCE_TO_NETWORK_CHAIN_TIP, NOT_SYNCED_ERROR_CODE,
+        },
+        get_block_template::{
+            calculate_transaction_roots, fake_coinbase_transaction, miner_fee,
+            standard_coinbase_outputs,
         },
         types::{
             default_roots::DefaultRoots, get_block_template::GetBlockTemplate,
@@ -50,6 +46,7 @@ use crate::methods::{
 
 pub mod config;
 pub mod constants;
+pub mod get_block_template;
 pub mod types;
 pub mod zip317;
 
@@ -181,8 +178,6 @@ where
         + 'static,
     SyncStatus: ChainSyncStatus + Clone + Send + Sync + 'static,
 {
-    // TODO: Add the other fields from the [`Rpc`] struct as-needed
-
     // Configuration
     //
     /// The configured network for this RPC service.
@@ -617,98 +612,4 @@ where
     }
 }
 
-// get_block_template support methods
-
-/// Returns the total miner fee for `mempool_txs`.
-pub fn miner_fee(mempool_txs: &[VerifiedUnminedTx]) -> Amount<NonNegative> {
-    let miner_fee: amount::Result<Amount<NonNegative>> =
-        mempool_txs.iter().map(|tx| tx.miner_fee).sum();
-
-    miner_fee.expect(
-        "invalid selected transactions: \
-         fees in a valid block can not be more than MAX_MONEY",
-    )
-}
-
-/// Returns the standard funding stream and miner reward transparent output scripts
-/// for `network`, `height` and `miner_fee`.
-///
-/// Only works for post-Canopy heights.
-pub fn standard_coinbase_outputs(
-    network: Network,
-    height: Height,
-    miner_address: transparent::Address,
-    miner_fee: Amount<NonNegative>,
-) -> Vec<(Amount<NonNegative>, transparent::Script)> {
-    let funding_streams = funding_stream_values(height, network)
-        .expect("funding stream value calculations are valid for reasonable chain heights");
-
-    let mut funding_streams: Vec<(Amount<NonNegative>, transparent::Address)> = funding_streams
-        .iter()
-        .map(|(receiver, amount)| (*amount, funding_stream_address(height, network, *receiver)))
-        .collect();
-    // The HashMap returns funding streams in an arbitrary order,
-    // but Zebra's snapshot tests expect the same order every time.
-    funding_streams.sort_by_key(|(amount, _address)| *amount);
-
-    let miner_reward = miner_subsidy(height, network)
-        .expect("reward calculations are valid for reasonable chain heights")
-        + miner_fee;
-    let miner_reward =
-        miner_reward.expect("reward calculations are valid for reasonable chain heights");
-
-    let mut coinbase_outputs = funding_streams;
-    coinbase_outputs.push((miner_reward, miner_address));
-
-    coinbase_outputs
-        .iter()
-        .map(|(amount, address)| (*amount, address.create_script_from_address()))
-        .collect()
-}
-
-/// Returns a fake coinbase transaction that can be used during transaction selection.
-///
-/// This avoids a data dependency loop involving the selected transactions, the miner fee,
-/// and the coinbase transaction.
-///
-/// This transaction's serialized size and sigops must be at least as large as the real coinbase
-/// transaction with the correct height and fee.
-fn fake_coinbase_transaction(
-    network: Network,
-    block_height: Height,
-    miner_address: transparent::Address,
-) -> TransactionTemplate<NegativeOrZero> {
-    // Block heights are encoded as variable-length (script) and `u32` (lock time, expiry height).
-    // They can also change the `u32` consensus branch id.
-    // We use the template height here, which has the correct byte length.
-    // https://zips.z.cash/protocol/protocol.pdf#txnconsensus
-    // https://github.com/zcash/zips/blob/main/zip-0203.rst#changes-for-nu5
-    //
-    // Transparent amounts are encoded as `i64`,
-    // so one zat has the same size as the real amount:
-    // https://developer.bitcoin.org/reference/transactions.html#txout-a-transaction-output
-    let miner_fee = 1.try_into().expect("amount is valid and non-negative");
-
-    let outputs = standard_coinbase_outputs(network, block_height, miner_address, miner_fee);
-    let coinbase_tx = Transaction::new_v5_coinbase(network, block_height, outputs).into();
-
-    TransactionTemplate::from_coinbase(&coinbase_tx, miner_fee)
-}
-
-/// Returns the transaction effecting and authorizing roots
-/// for `coinbase_tx` and `mempool_txs`.
-//
-// TODO: should this be spawned into a cryptographic operations pool?
-//       (it would only matter if there were a lot of small transactions in a block)
-pub fn calculate_transaction_roots(
-    coinbase_tx: &UnminedTx,
-    mempool_txs: &[VerifiedUnminedTx],
-) -> (merkle::Root, AuthDataRoot) {
-    let block_transactions =
-        || iter::once(coinbase_tx).chain(mempool_txs.iter().map(|tx| &tx.transaction));
-
-    let merkle_root = block_transactions().cloned().collect();
-    let auth_data_root = block_transactions().cloned().collect();
-
-    (merkle_root, auth_data_root)
-}
+// Put support functions in a submodule, to keep this file small.
